Lots of progress on high-level design of TN, node, edge classes
As a quick summary, TNs, nodes, and edges each get their own
classes, which are largely built on the graphs, nodes, and edges
defined in NetworkX. A lot of the functionality in these classes
is stored in the attributes of these corresponding objects in NX,
and one attribute is always the custom class which contains the
NX object in question.

This is a really simple design pattern, but it's shaping up to be
quite fruitful. For example, all of the tricky questions I had
regarding copy tensors, input tensors, and dangling edges were
all resolved by adding attributes to NX nodes that specify custom
behavior. It has also made it clear how I can have modular TNs,
where a node of a parent TN contains another whole TN with its own
parameters and whatnot.